add limits in tune menu when printing
limits:
  maximum printing speed
  maximum travel speed
  maximum extrusion speed
Can switch off with value=0, but it's still no good to change.
edit-settings-menu will revised later.
Put more text in one file (lower program memory, text easier to change).
Entry in MATERIAL.TXT changed (shorter) from bed_temperature_first_layer to bed_temp_first_layer.

+++ Marlin/Sd2CardExt.cpp
@@ -217,7 +217,7 @@ uint8_t Sd2CardExt::writeExtMask(uint8_t mio, uint8_t func,
   uint32_t arg =
       (((uint32_t)mio & 0x1) << 31) |
 	  (mio ? (((uint32_t)func & 0x7) << 28) : (((uint32_t)func & 0xF) << 27)) |
-	  (0x1 << 26) |
+	  ((uint32_t)0x1 << 26) |
 	  ((addr & 0x1FFFF) << 9) |
 	  mask;
